implemented partimaker system

+++ OpenGL_Framework/OpenGL_Framework/ParticleEffect.h
@@ -70,7 +70,7 @@ private:
 	ParticleList _Particles;
 	Texture _Texture;
 
-	float _Rate = 0.0f;
+	int _Rate = 0;
 	unsigned int _MaxParticles = 0;
 	unsigned int _NumCurrentParticles = 0;
